+++ R/redcap-read-oneshot.R
@@ -64,6 +64,9 @@
 #' specifies the maximum number of records to use for guessing column types.
 #' @param http_response_encoding  The encoding value passed to
 #' [httr::content()].  Defaults to 'UTF-8'.
+#' @param locale a [readr::locale()] object to specify preferences like
+#' number, date, and time formats.  This object is passed to
+#' [`readr::read_csv()`].  Defaults to [readr::default_locale()].
 #' @param verbose A boolean value indicating if `message`s should be printed
 #' to the R console during the operation.  The verbose output might contain
 #' sensitive information (*e.g.* PHI), so turn this off if the output might
@@ -180,6 +183,7 @@ redcap_read_oneshot <- function(
   guess_type                    = TRUE,
   guess_max                     = 1000L,
   http_response_encoding        = "UTF-8",
+  locale                        = readr::default_locale(),
   verbose                       = TRUE,
   config_options                = NULL
 ) {
@@ -210,6 +214,7 @@ redcap_read_oneshot <- function(
   checkmate::assert_integerish(guess_max                , any.missing=FALSE, len=1, lower=1)
   checkmate::assert_character(http_response_encoding    , any.missing=FALSE,     len=1)
 
+  checkmate::assert_class(    locale, "locale"          , null.ok = FALSE)
   checkmate::assert_logical(  verbose                   , any.missing=FALSE, len=1, null.ok=TRUE)
   checkmate::assert_list(     config_options            , any.missing=TRUE ,        null.ok=TRUE)
 
@@ -271,6 +276,7 @@ redcap_read_oneshot <- function(
           file            = I(kernel$raw_text),
           col_types       = col_types,
           guess_max       = guess_max,
+          locale          = locale,
           show_col_types  = FALSE
         ) %>%
         as.data.frame(),