+++ inst/test-data/decimal-comma/redcapr-specific/default-mismatched.R
@@ -1,0 +1,6 @@
+structure(list(record_id = c(1, 2, 3, 4), last_name = c("Valdez", 
+"Rich", "Furtado", "Akbar"), height_dot = c(1.54, 1.84, 1.95, 
+1.61), height_comma = c(154, 184, 195, 161), weight_dot = c(52.3, 
+92.3, 123.4, 45.9), weight_comma = c(523, 923, 1234, 459), bmi_dot = c(22.1, 
+27.3, 32.5, 17.7), bmi_comma = c(22.1, 27.3, 32.5, 17.7), demographics_complete = c(0, 
+0, 0, 0)), row.names = c(NA, -4L), class = "data.frame")